pyproject.toml
[tool.poetry]
name = "imgsys"
version = "0.1.0"
description = "Image Classification system - -Phase 0 scaffolding"
authors = ["Devenair Thomas <deventhomas55@gmail.com>"]
readme = "README.md"
packages = [{ include = "imgsys" }]

[tool.poetry.dependencies]
python = "^3.10.6"
pydantic = "^2.8"
jsonschema = "^4.22"
Pillow = "^10.4"
piexif = "^1.1.3"
typer = "^0.12.3"
PyYAML = "^6.0.2"
rich = "^13.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
black = "^24.4"
ruff = "^0.5"
